* changes:
  Make Process.getFreeMemory() return a long instead of an int, to better work on the simulator now, and phones 2 years from now.

+++ core/java/android/os/Process.java
@@ -739,7 +739,7 @@ public class Process {
     public static final native void sendSignal(int pid, int signal);
     
     /** @hide */
-    public static final native int getFreeMemory();
+    public static final native long getFreeMemory();
     
     /** @hide */
     public static final native void readProcLines(String path,
